updates so docker can be run from a script and if the option not given then run all the tags by default
