+++ environment/Configuration.py
@@ -1,1 +1,2 @@
 ELEMENT_WAIT_TIME = 5  # time in seconds
+MAX_PARALLEL_SCENARIOS = 5  # currently no more than 5
